requirements.txt
selenium>=4.10.0
webdriver-manager>=4.0.1
python-dotenv>=1.0.1
pandas>=2.2.1
